requirements.txt
google-cloud-speech
spacy
